+++ src/test/kotlin/giga/LocalRegexClassifierTest.kt
@@ -25,14 +25,14 @@ class LocalRegexClassifierTest {
     @Test
     fun `classifies coder creation`() = runBlocking {
         val classifier = LocalRegexClassifier()
-        val category = classifier.classify(body("создай файл"))
+        val category = classifier.classify(body("Кодер, поправь ридми"))
         assertEquals(ToolCategory.CODER, category)
     }
 
     @Test
     fun `classifies coder read`() = runBlocking {
         val classifier = LocalRegexClassifier()
-        val category = classifier.classify(body("прочитай readme и используй grep"))
+        val category = classifier.classify(body("Можешь поправить баги в тестах?"))
         assertEquals(ToolCategory.CODER, category)
     }
 
@@ -46,21 +46,28 @@ class LocalRegexClassifierTest {
     @Test
     fun `classifies browser tabs`() = runBlocking {
         val classifier = LocalRegexClassifier()
-        val category = classifier.classify(body("переключи вкладку браузера"))
+        val category = classifier.classify(body("Открой только что закрытую вкладку"))
         assertEquals(ToolCategory.BROWSER, category)
     }
 
     @Test
-    fun `classifies config volume`() = runBlocking {
+    fun `classifies config instruction`() = runBlocking {
         val classifier = LocalRegexClassifier()
-        val category = classifier.classify(body("настрой громкость"))
+        val category = classifier.classify(body("Сохрани инструкцию: когда я говорю TG Шам, отправь Шамилю сообщение в телеграме со словом Привет"))
+        assertEquals(ToolCategory.CONFIG, category)
+    }
+
+    @Test
+    fun `classifies config volume and instruction`() = runBlocking {
+        val classifier = LocalRegexClassifier()
+        val category = classifier.classify(body("Запомни инструкцию: когда я говорю «Ускорь», ускорь скорость речь на 40 слов в минуту"))
         assertEquals(ToolCategory.CONFIG, category)
     }
 
     @Test
     fun `classifies config speed`() = runBlocking {
         val classifier = LocalRegexClassifier()
-        val category = classifier.classify(body("config speed"))
+        val category = classifier.classify(body("Замедли скорость речи"))
         assertEquals(ToolCategory.CONFIG, category)
     }
 
@@ -74,14 +81,14 @@ class LocalRegexClassifierTest {
     @Test
     fun `classifies io download`() = runBlocking {
         val classifier = LocalRegexClassifier()
-        val category = classifier.classify(body("скачай файл"))
+        val category = classifier.classify(body("скачай файл с только что отредактированным pdf файлом"))
         assertEquals(ToolCategory.IO, category)
     }
 
     @Test
     fun `classifies io screenshot`() = runBlocking {
         val classifier = LocalRegexClassifier()
-        val category = classifier.classify(body("сделай скриншот"))
+        val category = classifier.classify(body("Что на экране?"))
         assertEquals(ToolCategory.IO, category)
     }
